feat(spectator): refactor team box component and move to separate file
